pipelines: run crates io fetch concurrently

+++ fpr/db.py
@@ -1,7 +1,7 @@
 import contextlib
 import json
 import sqlite3
-from typing import Iterator
+from typing import Iterator, Dict
 
 
 @contextlib.contextmanager
@@ -13,8 +13,9 @@ def connect(db_file: str) -> Iterator[sqlite3.Connection]:
         conn.close()
 
 
-def create_crates_io_meta_table(cursor: sqlite3.Cursor):
-    cursor.execute("DROP TABLE IF EXISTS crates_io_metadata")
+def create_crates_io_meta_table(cursor: sqlite3.Cursor, drop_if_exists: bool = False):
+    if drop_if_exists:
+        cursor.execute("DROP TABLE IF EXISTS crates_io_metadata")
     cursor.execute(
         "CREATE TABLE crates_io_metadata(id INTEGER PRIMARY KEY AUTOINCREMENT, crate_meta JSON NOT NULL)"
     )
@@ -28,7 +29,8 @@ def crate_name_in_db(cursor: sqlite3.Cursor, crate_name: str) -> bool:
     return cursor.fetchone() is not None
 
 
-def save_crate_meta(cursor: sqlite3.Cursor, json_str: str):
+def save_crate_meta(cursor: sqlite3.Cursor, json_dict: Dict):
     cursor.execute(
-        "INSERT INTO crates_io_metadata(crate_meta) VALUES (?)", (json.dumps(json_str),)
+        "INSERT INTO crates_io_metadata(crate_meta) VALUES (?)",
+        (json.dumps(json_dict),),
     )
